registry fix
