+++ catalyst/registry/registry.py
@@ -1,5 +1,3 @@
-# flake8: noqa
-# @TODO: code formatting issue for 20.07 release
 """
 Registry.
 .. todo:: Representative docstring for this module
