+++ catalyst/utils/metrics/tests/test_average_precision.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 import math
 
 import torch
